requirements.txt
google-genai
streamlit

#just run pip install -r requirements.txt

